Ajuste en tipo de dato en modelo de respuesta y controlador de Perfil Físico del Usuario

+++ Kilometros WebAPI/Models/HttpGet/My_Controllers/PhysiqueResponse.cs
@@ -10,7 +10,7 @@ namespace Kilometros_WebAPI.Models.HttpGet.My_Controllers {
         public int Weight;
         public string Sex;
 
-        public short StrideLengthWalking;
-        public short StrideLengthRunning;
+        public int StrideLengthWalking;
+        public int StrideLengthRunning;
     }
 }
